requirements.txt
xmpppy==0.7.1
python-dotenv==1.0.0
phonenumbers==8.13.7
